+++ src/lib/pattern/Airport.ts
@@ -76,7 +76,7 @@ export class Airport {
     this.magneticDeclination = airportNormalized.magdec;
 
     /**
-     * @type {number} a tome zone which only considers the longitude, rounded to the full hour, in hours difference to UTC
+     * @type {number} a time zone which only considers the longitude, rounded to the full hour, in hours difference to UTC
      * @see https://en.wikipedia.org/wiki/Nautical_time
      */
     this.nauticalTimezone = Math.round(this.position.longitude / 15);