+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added additional AviationWeather API methods
 - Improved pattern generator to also work for instrument approach procedures
+- Added holding pattern generator
 
 ## [2.6.1]